refactor: move routes
