+++ frontend/svelte/src/routes/AuthPage.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import {onDestroy} from 'svelte';
   import type {Unsubscriber} from 'svelte/types/runtime/store';
-  import {AuthStore, authStore} from './lib/stores/auth.store';
+  import {AuthStore, authStore} from '../lib/stores/auth.store';
 
   let signedIn: boolean = false;
