+++ frontend/svelte/src/routes/CanistersPage.svelte
