+++ frontend/svelte/src/routes/ContentPage.svelte
