+++ frontend/svelte/src/routes/NeuronsPage.svelte
