+++ frontend/svelte/src/routes/VotingPage.svelte
